+++ lib/amazonka/CHANGELOG.md
@@ -97,6 +97,8 @@ Released: **?**, Compare: [2.0.0-rc1](https://github.com/brendanhay/amazonka/com
 
 ### Fixed
 
+- `amazonka-core`: Correctly double-url-encode request paths when computing V4 signatures
+[\#812](https://github.com/brendanhay/amazonka/pull/812)
 - `amazonka-core`: Correctly compute body length for `hashedFileRange`
 [\#794](https://github.com/brendanhay/amazonka/pull/794)
 - Generator: Correctly generate `ToJSON` instances for requests which set a `"payload":` field in `"type": "structure"` definitions.